pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "begone-convert"
authors = [{name = "Daniel Rocha", email = "daniel@rocha.io"}]
dynamic = ["version", "description"]
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
    "pyyaml >=6.0.2",
    "typer >=0.12.5",
]

[project.scripts]
begone-convert = "begone_convert:convert"

[tool.ruff]
preview = true

[tool.ruff.lint]
extend-select = [
  # "D",   # pydocstyle
  "I",   # isort
  "UP",  # pyupgrade
  "S",   # flake8-bandit
  "B",   # flake8-bugbear
  "COM", # flake8-commas
  "C4",  # flake8-comprehensions
  "ISC", # flake8-implicit-str-concat
  "FIX", # flake8-fixme
  "FA",  # flake8-future-annotations
  "SIM", # flake8-simplify
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "D100", "D103", "D104"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_untyped_defs = true
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
warn_unused_configs = true
